pyproject.toml
[project]
name = "langchain-cache-gcs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Chengyueh Sha", email = "chengyueh.sha@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.1.1",
    "langchain>=0.3.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "langchain-xai>=0.2.4",
    "python-dotenv>=1.1.1",
]
